+++ CHANGES.md
@@ -13,6 +13,7 @@ Variables that cannot be normalized to use these dimensions will not be shown.
 If `normalize_data` is False, dimensions will not be changed and all data variables will be shown 
 (i.e., all variables that are not dimensionns, that are numeric and that have more than one dimension).
 * Updated setup.py [#16]()https://github.com/dcs4cop/xcube-cci/issues/16)
+* Added opener parameters `time_range` and `spatial_res`
 
 ## Changes in 0.5.0.